Enlèvement des methodes estCamoufle et estVolant, rajout d'instance of Remplacement de int dans les x et y des bâtiments par des double Ajout de methodes pour faciliter le code et l'amélioration des bâtiments. Ajout d'un attribut niveauMax.

+++ src/main/java/universite_paris8/iut/mcheema/codesource/modele/ennemi/Bogue.java
@@ -1,14 +1,14 @@
 package universite_paris8.iut.mcheema.codesource.modele.ennemi;
 
-/**
- * Bogue qui étend la classe Ennemi, contient les mêmes attributs que Ennemi.
- */
 
 import universite_paris8.iut.mcheema.codesource.modele.Environnement;
 import universite_paris8.iut.mcheema.codesource.modele.Point;
 import java.util.ArrayList;
 
+/**
+ * Bogue qui étend la classe Ennemi, contient les mêmes attributs que Ennemi.
+ */
 public class Bogue extends Ennemi {
     public Bogue(Environnement env, ArrayList<Point> chemin) {super(2, 2, 2, env, chemin);}
-    public Bogue(int x, int y, Environnement env, ArrayList<Point> chemin) {super(2, 5, 2, env, chemin);}
+    public Bogue(double x, double y, Environnement env, ArrayList<Point> chemin) {super(2, 5, 2, env, chemin);}
 }
